refactor: create a directory for error files
