+++ packages/rest-api-client/src/error/__tests__/KintoneAllRecordsError.test.ts
@@ -1,6 +1,6 @@
 import { KintoneAllRecordsError } from "../KintoneAllRecordsError";
 import { KintoneRestAPIError } from "../KintoneRestAPIError";
-import { ErrorResponse } from "../http/HttpClientInterface";
+import { ErrorResponse } from "../../http/HttpClientInterface";
 
 describe("KintoneAllRecordsError", () => {
   let kintoneAllRecordsError: KintoneAllRecordsError;
