+++ packages/rest-api-client/src/error/__tests__/KintoneRestAPIError.test.ts
